+++ models/staging/camara_deputados/stg_camara_legislaturas.sql
@@ -9,8 +9,8 @@ WITH source AS (
 
 renamed AS (
     SELECT
-        idlegislatura                       AS legislatura_id_nk,
-        id                                  AS deputado_id_fk,
+        idlegislatura::BIGINT               AS legislatura_id_nk,
+        id::BIGINT                          AS deputado_id_fk,
         nome,
         siglauf                             AS uf,
         SPLIT_PART(uripartido, '/', 7)::INT AS partido_id_fk
